+++ mptest/lib/t_skip.m
@@ -1,9 +1,9 @@
 function t_skip(cnt, msg)
 %T_SKIP  Skips a number of tests.
 %   T_SKIP(CNT, MSG) increments the global test count and skipped tests
-%   count. Prints 'skipped tests x..y : ' followed by the MSG, unless the
-%   global variable t_quiet is true. Intended to be called between calls to
-%   T_BEGIN and T_END.
+%   count. Prints 'skipped x..y : ' followed by the MSG, unless
+%   T_BEGIN was called with input QUIET equal true. Intended to be
+%   called between calls to T_BEGIN and T_END.
 %
 %   Example:
 %       quiet = 0;
@@ -17,7 +17,7 @@ function t_skip(cnt, msg)
 
 
 %   MP-Test
-%   Copyright (c) 2004-2016, Power Systems Engineering Research Center (PSERC)
+%   Copyright (c) 2004-2020, Power Systems Engineering Research Center (PSERC)
 %   by Ray Zimmerman, PSERC Cornell
 %
 %   This file is part of MP-Test.